Create Blood Donation System
Blood Donation System is a project which aims at bridging the gap between blood donors and those in need of the live saving blood.  It serves as a platform that connects donors with blood banks and recipients, there by developing a culture of voluntary blood donation to address the issue of critical shortage of blood supply. 

This system involves a user interface, the system allows individuals to register as donors and provide essential information about their blood type, contact details, and availability for donation. Blood banks and medical facilities can access this information to identify suitable donors in times of emergencies or critical shortage.